api/autoworx_service/requirements.txt
python-dotenv>=1.0.1,<1.1.0
Django>=5.0.6,<5.1
djangorestframework>=3.15.1,<3.16
psycopg2>=2.9.9,<3
drf-spectacular>=0.27.2,<0.28
django-cors-headers>=4.4.0,<4.5.0
pillow>=10.4.0,<10.5
drf-yasg>=1.21.7,<1.22